# Vectorized the association assets <-> sites