--- a/debian/changelog
+++ b/debian/changelog
@@ -1,4 +1,5 @@
   [Michele Simionato]
+  * Vectorized the association assets <-> hazard sites (up to 15x speedup)
   * Fixed bug in disaggregation calculations associated to an error
     in the line `[mag] = np.unique(np.round(ctx.mag, 6))`
   * Optimized the calculation of geohashes by using numba