--- a/openquake/commonlib/readinput.py
+++ b/openquake/commonlib/readinput.py
@@ -999,9 +999,9 @@ def get_sitecol_assetcol(oqparam, haz_sitecol=None, exp_types=()):
     if haz_sitecol.mesh != exp.mesh:
         # associate the assets to the hazard sites
         t0 = time.time()
-        sitecol, assets_by, discarded = geo.utils.assoc(
-            exp.assets_by_site, haz_sitecol, haz_distance,
-            'filter')
+        sitecol, assets_by, discarded = geo.utils._GeographicObjects(
+            haz_sitecol).assoc2(exp.mesh, exp.assets_by_site, haz_distance,
+                                'filter')
         num_assets = sum(len(assets) for assets in assets_by)
         dt = time.time() - t0
         logging.info('Associated {:_d} assets to {:_d} sites in {:.1f}s'.