With a 32x speedup in the USA model:
```
# master
| calc_57449               | time_sec | memory_mb | counts |
|--------------------------+----------+-----------+--------|
| EventBasedCalculator.run | 196.1    | 1_312     | 1      |
| importing inputs         | 192.1    | 3_195     | 1      |
| reading exposure         | 125.9    | 615.2     | 1      |
Associated 4_512_042 assets to 49_671 sites in 94.8s

# asset-site
| calc_57448               | time_sec | memory_mb | counts |
|--------------------------+----------+-----------+--------|
| EventBasedCalculator.run | 105.1    | 1_387     | 1      |
| importing inputs         | 101.0    | 3_266     | 1      |
| reading exposure         | 34.8     | 614.1     | 1      |
Associated 4_512_042 assets to 49_671 sites in 3.0s
```
We can import the USA exposure on davis in 17 minutes now (was 32 minutes) because now the association of
22_551_641 assets to 55_771 sites takes an incredible 16.7s.